% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertemp-package.r
\docType{package}
\name{convertemp}
\alias{convertemp}
\alias{convertemp-package}
\title{convertemp: Easily convert temperatures}
\description{
Easily convert temperatures between any combination of Celsius,
Kelvin, and Fahrenheit.
}

